third_party/blink/web_tests/platform/generic/fast/box-sizing/table-cell-expected.txt

Tests that display: table-cell and box-sizing: border-box work when used together.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

div-display-table-cell
PASS element.offsetWidth is 80
PASS element.offsetHeight is 30
td
PASS element.offsetWidth is 80
PASS element.offsetHeight is 30
PASS successfullyParsed is true

TEST COMPLETE
